% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Csimilar}
\alias{Csimilar}
\title{基于俩文本Character之间的编辑距离，计算文本相似度}
\usage{
Csimilar(a, b, methd = "lv")
}
\arguments{
\item{a}{文本a}

\item{b}{文本b}

\item{methd}{方法，默认lv}
}
\value{
返回文本相似度数据
}
\description{
基于俩文本Character之间的编辑距离，计算文本相似度
}
