% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deDups}
\alias{deDups}
\title{对指定字段的数据进行去重操作，deDuplication,deDups}
\usage{
deDups(dt, rpcols = NULL, dwcol = NULL, ori = 1)
}
\arguments{
\item{dt}{待去重数据，可为data.frame，data.table以及list类型数据}

\item{rpcols}{repeat coloumns,为需去重的列,存在多个相同值，可为多列。}

\item{dwcol}{去重依据列,可为id、日期时间类，取值唯一。}

\item{ori}{为去重方向，1为保留dwcol中的最大值，0为保留dwcol中的最小值}
}
\value{
返回去重后的data.table数据
}
\description{
如果rpcols对应的dwcol包含有效值与NA，NA会全部保留,

如果rpcols对应的dwcol全部为NA，不处理(仅做全部列的unique处理)
}
