% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downa}
\alias{downa}
\title{下载文件到指定地址，提供文件链接及对应的文件名}
\usage{
downa(url, dest = "./", dw = 0)
}
\arguments{
\item{url}{待下载文件的URL地址}

\item{dest}{文件下载地址，包含文件名}

\item{dw}{存在同名文件是否下载，默认为0不下载}
}
\value{
下载成功返回地址，下载失败返回0
}
\description{
下载文件到指定地址，提供文件链接及对应的文件名
}
