% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAbsr}
\alias{getAbsr}
\title{根据项目当前位置获取参数的绝对地址,get absolute address
若地址开头为 * 则获取icloud云盘地址}
\usage{
getAbsr(adr)
}
\arguments{
\item{adr}{文件地址}
}
\value{
属性credb标识adr是否存在
}
\description{
根据项目当前位置获取参数的绝对地址,get absolute address
若地址开头为 * 则获取icloud云盘地址
}
