% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCloudr}
\alias{getCloudr}
\title{get icloud address}
\usage{
getCloudr()
}
\value{
NULL OR address of icloud folder.
}
\description{
depends OS system : windows or mac
}
\details{
direction in windows:C:/Users/Administrator/iCloudDrive

direction in macos:/Users/longsa/Library/Mobile Documents/com~apple~CloudDocs
}
