% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDtag}
\alias{getDtag}
\title{根据开始日期，以及给定日期，获取从基准日期开始的第x周/月}
\usage{
getDtag(rdate = NULL, sdate = ymd(20250101), width = "week")
}
\arguments{
\item{rdate}{给定的待获取的基于基准日期开始计算的日期}

\item{sdate}{基准开始日期}

\item{width}{计算间隔，可以是day/week/month}
}
\value{
返回是第几day/week/month
}
\description{
根据开始日期，以及给定日期，获取从基准日期开始的第x周/月
}
