% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pv.annu.growth}
\alias{pv.annu.growth}
\title{固定增长率的年金现值,pmt(payment for the first period)}
\usage{
pv.annu.growth(r, g, n, pmt = 1)
}
\arguments{
\item{r}{折现率}

\item{g}{增长率}

\item{n}{期数}

\item{pmt}{payment for the first period}
}
\value{
返回年金现值
}
\description{
固定增长率的年金现值,pmt(payment for the first period)
}
