% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replas}
\alias{replas}
\title{根据patterns-reps对进行文本替换，patterns可为数字串
如x为数值型，默认增加|转换为字符，且reps前后增加|的匹配
如|234|23|，替换23时，不替换｜234｜中的23.}
\usage{
replas(x, pats, reps)
}
\arguments{
\item{x}{待进行文本替换的字符或者数字}

\item{pats}{待替换的文本/数字串}

\item{reps}{替换后的文本串；与pats一一对应}
}
\value{
返回替换后的文本串
}
\description{
根据patterns-reps对进行文本替换，patterns可为数字串
如x为数值型，默认增加|转换为字符，且reps前后增加|的匹配
如|234|23|，替换23时，不替换｜234｜中的23.
}
