% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundt}
\alias{roundt}
\title{指定小数点位数，并且丢弃指定位数后的数字
对plotly的使用十分重要}
\usage{
roundt(x, digit = 2)
}
\arguments{
\item{x}{待处理的数字}

\item{digit}{小数点位数}
}
\value{
返回double类型数据，小数点位数digit
}
\description{
指定小数点位数，并且丢弃指定位数后的数字
对plotly的使用十分重要
}
