% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDigit}
\alias{setDigit}
\title{将dt中指定列的数值转换为指定长度单位}
\usage{
setDigit(dt, mcols = "value", digit = 1, skipx = list())
}
\arguments{
\item{dt}{data.table数据}

\item{mcols}{为需要进行转换的列}

\item{digit}{进行转换的单位基数，默认为1，不进行转换}

\item{skipx}{为“名-值”对，表示此列对应值所在行的mcols列的值不进行转换}
}
\value{
返回转换后的数据
}
\description{
将dt中指定列的数值转换为指定长度单位
}
