% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Tsimilar}
\alias{Tsimilar}
\title{计算俩片段间重叠的时间time长度与第一个片段长度之比}
\usage{
Tsimilar(sega, segb, methd = "x")
}
\arguments{
\item{sega}{片段a的开始-结束时间}

\item{segb}{片段b的开始-结束时间}

\item{methd}{选择比较方式，暂时没有！}
}
\value{
返回相似度数据
}
\description{
计算俩片段间重叠的时间time长度与第一个片段长度之比
}
